chore(lint): resolve all frontend lint warnings + errors (refs #212)
- Extract formatError out of ErrorMessage.tsx to satisfy
  react-refresh/only-export-components.
- AlbumsPage: replace view-change reset effect with the render-phase
  prev-value pattern (avoids react-hooks/set-state-in-effect).
- DevicePicker: move device fetch out of useEffect into the toggle
  handler; request-id ref handles cancellation.
- PlayerBar: justify-and-suppress two exhaustive-deps warnings with
  inline rationale tied to the #185 streamUrl-salt hazard.
- AGENTS.md per-task checklist: require pnpm lint to exit with zero
  errors and zero warnings before declaring done.
