+++ frontend/src/components/ui/ErrorMessage.tsx
@@ -1,13 +1,5 @@
 import { AlertCircle } from "lucide-react";
-import { SubsonicError } from "@/lib/subsonic";
-
-export function formatError(err: unknown): { title: string; code?: number } {
-  if (err instanceof SubsonicError) {
-    return { title: err.message || "Request failed", code: err.code };
-  }
-  if (err instanceof Error) return { title: err.message };
-  return { title: "Unknown error" };
-}
+import { formatError } from "./formatError";
 
 export function ErrorMessage({ error }: { error: unknown }) {
   const { title, code } = formatError(error);
